common: factor out common c++ version of 'which'
